+++ internal/app/shell.go
@@ -255,6 +255,7 @@ func Shell(props ShellProps) uic.Node {
 		uic.CreateElement(AddHost),
 		uic.CreateElement(TxnEditHost),
 		uic.CreateElement(TxnColumnsHost),
+		uic.CreateElement(TxnSmartCatHost),
 		uic.CreateElement(AccountEditHost),
 		uic.CreateElement(BudgetEditHost),
 		uic.CreateElement(GoalEditHost),